+++ test/clojure/nginx/clojure/test_all.clj
@@ -557,7 +557,7 @@
              (is (= 200 (:status r)))
              (is (or (= "hello,a!/mytestpath" (:body r)) (= "hello,b!/mytestpath" (:body r))))))
        (testing "rewrite proxy pass by body -- small body"
-           (let [r (client/post (str "http://" *host* ":" *port* "/javarewritebybodyproxy/") {:follow-redirects false, :body "ub"})
+           (let [r (client/post (str "http://" *host* ":" *port* "/javarewritebybodyproxy/") {:follow-redirects false, :body "ub" :throw-exceptions false :socket-timeout 10000})
                  h (:headers r)
                  b (r :body)]
              (debug-println r)
@@ -565,7 +565,7 @@
              (is (= 200 (:status r)))
              (is  (= "hello,b!/javarewritebybodyproxy/" (:body r)) )))
         (testing "rewrite proxy pass by body -- large body"
-           (let [r (client/post (str "http://" *host* ":" *port* "/javarewritebybodyproxy/") {:follow-redirects false, :body (clojure.java.io/file "test/nginx-working-dir/post-rewrite-large-body-data")})
+           (let [r (client/post (str "http://" *host* ":" *port* "/javarewritebybodyproxy/") {:follow-redirects false, :body (clojure.java.io/file "test/nginx-working-dir/post-rewrite-large-body-data") :throw-exceptions false :socket-timeout 10000})
                  h (:headers r)
                  b (r :body)]
              (debug-println r)